+++ packages/shared/components/Button.tsx
@@ -1,6 +1,6 @@
+import twMerge from "@digitalcheck/shared/utils/tailwindMerge";
 import { cloneElement, type ReactElement } from "react";
 import { Link } from "react-router-dom";
-import twMerge from "utils/tailwind-merge";
 
 type Props = {
   text?: string;
