+++ packages/shared/components/Container.tsx
@@ -1,5 +1,5 @@
+import twMerge from "@digitalcheck/shared/utils/tailwindMerge";
 import type { PropsWithChildren } from "react";
-import twMerge from "utils/tailwind-merge";
 import { BACKGROUND_COLORS, type BackgroundColor } from ".";
 
 type ContainerProps = {
