+++ packages/shared/components/Heading.tsx
@@ -1,6 +1,6 @@
 import RichText from "@digitalcheck/shared/components/RichText";
+import twMerge from "@digitalcheck/shared/utils/tailwindMerge";
 import type { ReactNode } from "react";
-import twMerge from "utils/tailwind-merge";
 
 export type HeadingProps = {
   tagName?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p" | "div";
