+++ packages/shared/components/Question.tsx
@@ -1,5 +1,5 @@
+import twMerge from "@digitalcheck/shared/utils/tailwindMerge";
 import { PropsWithChildren } from "react";
-import twMerge from "utils/tailwind-merge";
 import { BackgroundColor } from ".";
 import Box, { BoxProps } from "./Box";
 import DetailsSummary from "./DetailsSummary";
